pyproject.toml
[tool.poetry]
name = "kueue"
version = "0.1.0"
description = "Distributed Task Queue - using Kubernetes and Kafka"
authors = ["Jose Rojas <jose.rojas95@mail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
confluent-kafka = "^1.5.0"
pydantic = "^1.7.3"
wrapt = "^1.12.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = "^5.3.0"
flakehell = "^0.7.1"
flake8 = "^3.8.4"
flake8-black = "^0.2.1"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.3.0"
flake8-isort = "^4.0.0"
flake8-mutable = "^1.2.0"
flake8-print = "^4.0.0"
hostsman = "^1.1.5"
isort = "^5.6.4"
pytest-kind = "^20.10.0"
pytest-cov = "^2.10.1"
pytest-html = "^3.1.0"
pytest-mock = "^3.3.1"
requests = "^2.25.0"
ipython = "*"
debugpy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flakehell]
# make output nice
format = "colored"
# show line of source code in output
show_source = true
max_line_length = 100

[tool.flakehell.plugins]
flake8-black = ["+*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-print = ["+*"]
mccabe = ["+*"]
pycodestyle = [
    "+*",
    "-W503"  # black compat: line break before binary operator
]
pyflakes = ["+*"]
pylint = ["+*"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
line_length = 100
multi_line_output = 3  # Matches black formatting
use_parentheses = true
include_trailing_comma = true
force_grid_wrap = 0
ensure_newline_before_comments = true

[tool.pytest.ini_options]
junit_suite_name = "kueue"
junit_family = "xunit2"
markers = [
    "integration: integration tests."
]
addopts = """
    --keep-cluster
    --cov=kueue \
    --cov-report=html:tests/coverage \
    --cov-report=term \
    --html=tests/results/index.html \
    --junitxml=tests/results/summary.xml \
    -vv
"""

[tool.coverage.run]
omit = ["tests"]
